[TASK] Refactor additional buttons
