+++ Classes/Utility/ProductUtility.php
@@ -70,20 +70,6 @@ class ProductUtility
     const ORDER_STATE_COOKIE_NAME = 'pxa_pm_order_state';
 
     /**
-     * @var Dispatcher
-     */
-    public static $signalSlotDispatcher;
-
-    /**
-     * @return Dispatcher
-     */
-    public static function getSignalSlotDispatcher(): Dispatcher
-    {
-        self::$signalSlotDispatcher = self::$signalSlotDispatcher ?? GeneralUtility::makeInstance(Dispatcher::class);
-        return self::$signalSlotDispatcher;
-    }
-
-    /**
      * Get array of uids of categories for product
      *
      * @param $product
@@ -461,54 +447,5 @@ class ProductUtility
         );
 
         return $productsQuantityData;
-    }
-
-    /**
-     * @param Product $product
-     * @param array $buttons
-     * @return array
-     */
-    public static function getAdditionalButtons(Product $product, array $buttons = []) : array
-    {
-        /**
-         * Generate additional buttons
-         * Should an array that follows this structure
-         * [
-         *     [
-         *         'name' => 'Do something',
-         *         'link' => 'https://www.example.com',
-         *         'classes' => [],
-         *         'order' => '100'
-         *     ],
-         *     [
-         *         'name' => 'Buy',
-         *         'link' => 'https://www.test.net',
-         *         'classes' => ['beauty', 'clarence'],
-         *         'order' => '20'
-         *     ],
-         * ]
-         *
-         * name - button text
-         * link - button link
-         * classes - array of additional button classes
-         * order - used to specify buttons order
-         */
-
-        // Add a signal slot so other extension could add additional buttons
-        $dispatcher = self::getSignalSlotDispatcher();
-        $dispatcher->dispatch(__CLASS__, 'BeforeProcessingAdditionalButtons', [$product, &$buttons]);
-
-        // Process
-        foreach ($buttons as &$button) {
-            $button['classes'] = empty($button['classes']) ? '' : implode(' ', $button['classes']);
-        }
-        unset($button);
-
-        // Sort
-        usort($buttons, function ($a, $b) {
-            return $a['order'] - $b['order'];
-        });
-
-        return $buttons;
     }
 }
