--- Resources/Private/Templates/Standard/SingleViewAdditionalButtons.html
@@ -1,9 +1,0 @@
-<html xmlns:f="http://typo3.org/ns/TYPO3/CMS/Fluid/ViewHelpers" data-namespace-typo3-fluid="true">
-
-<f:for each="{buttons}" as="button">
-    <f:spaceless>
-        <a href="{button.link}" class="btn btn-default product-header-btn">
-            {button.name}
-        </a>
-    </f:spaceless>
-</f:for>
