+++ Tests/Unit/Utility/ProductUtilityTest.php
@@ -300,170 +300,4 @@ class ProductUtilityTest extends UnitTestCase
 
         $this->assertEquals($expect, ProductUtility::calculateOrderTotalTax($order));
     }
-
-    /**
-     * @test
-     */
-    public function getAdditionalButtonsIsAlteringButtonUsingSignalSlots()
-    {
-        $expected = [
-            [
-                'name' => 'Sell me',
-                'link' => 'http://example.sell.com',
-                'classes' => '',
-                'order' => 20
-            ],
-            [
-                'name' => 'Buy me',
-                'link' => 'http://example.com',
-                'classes' => '',
-                'order' => 120
-            ]
-        ];
-
-
-        $signalSlotDispatcherMock = $this->getAccessibleMock(
-            Dispatcher::class,
-            ['dispatch']
-        );
-
-        $buttons = [];
-        $product = $this->getAccessibleMock(Product::class, ['dummy']);
-
-        $signalSlotDispatcherMock->expects($this->once())
-            ->method('dispatch')
-            ->with(ProductUtility::class, 'BeforeProcessingAdditionalButtons', [$product, $buttons])
-            ->will($this->returnCallback(function ($class, $name, $params) use ($expected) {
-                $params[1] = [
-                    [
-                        'name' => 'Sell me',
-                        'link' => 'http://example.sell.com',
-                        'classes' => [],
-                        'order' => 20
-                    ],
-                    [
-                        'name' => 'Buy me',
-                        'link' => 'http://example.com',
-                        'classes' => [],
-                        'order' => 120
-                    ]
-                ];
-            }));
-
-        ProductUtility::$signalSlotDispatcher = $signalSlotDispatcherMock;
-        $result = ProductUtility::getAdditionalButtons($product, $buttons);
-
-        $this->assertEquals($expected, $result);
-    }
-
-    /**
-     * @test
-     */
-    public function getAdditionalButtonsSortsTheButtonsAccordingToOrderField()
-    {
-        $expected = [
-            [
-                'name' => 'Sell me',
-                'link' => 'http://example.sell.com',
-                'classes' => '',
-                'order' => 20
-            ],
-            [
-                'name' => 'Buy me',
-                'link' => 'http://example.com',
-                'classes' => '',
-                'order' => 120
-            ]
-        ];
-
-
-        $signalSlotDispatcherMock = $this->getAccessibleMock(
-            Dispatcher::class,
-            ['dispatch']
-        );
-
-        $buttons = [];
-        $product = $this->getAccessibleMock(Product::class, ['dummy']);
-
-        $signalSlotDispatcherMock->expects($this->once())
-            ->method('dispatch')
-            ->with(ProductUtility::class, 'BeforeProcessingAdditionalButtons', [$product, $buttons])
-            ->will($this->returnCallback(function ($class, $name, $params) use ($expected) {
-                $params[1] = [
-                    [
-                        'name' => 'Buy me',
-                        'link' => 'http://example.com',
-                        'classes' => [],
-                        'order' => 120
-                    ],
-                    [
-                        'name' => 'Sell me',
-                        'link' => 'http://example.sell.com',
-                        'classes' => [],
-                        'order' => 20
-                    ]
-                ];
-            }));
-
-        ProductUtility::$signalSlotDispatcher = $signalSlotDispatcherMock;
-        $result = ProductUtility::getAdditionalButtons($product, $buttons);
-
-        $this->assertEquals($expected, $result);
-    }
-
-    /**
-     * @test
-     */
-    public function getAdditionalButtonsConvertsClassesListToAClassString()
-    {
-        $expected = [
-            [
-                'name' => 'Sell me',
-                'link' => 'http://example.sell.com',
-                'classes' => 'testclass dummy2',
-                'order' => 20
-            ],
-            [
-                'name' => 'Buy me',
-                'link' => 'http://example.com',
-                'classes' => 'testclass',
-                'order' => 120
-            ]
-        ];
-
-
-        $signalSlotDispatcherMock = $this->getAccessibleMock(
-            Dispatcher::class,
-            ['dispatch']
-        );
-
-        $buttons = [];
-
-        $product = $this->getAccessibleMock(Product::class, ['dummy']);
-
-        $signalSlotDispatcherMock->expects($this->once())
-            ->method('dispatch')
-            ->with(ProductUtility::class, 'BeforeProcessingAdditionalButtons', [$product, $buttons])
-            ->will($this->returnCallback(function ($class, $name, $params) use ($expected) {
-                $params[1] = [
-                    [
-                        'name' => 'Sell me',
-                        'link' => 'http://example.sell.com',
-                        'classes' => ['testclass', 'dummy2'],
-                        'order' => 20
-                    ],
-                    [
-                        'name' => 'Buy me',
-                        'link' => 'http://example.com',
-                        'classes' => ['testclass'],
-                        'order' => 120
-                    ]
-                ];
-            }));
-
-        ProductUtility::$signalSlotDispatcher = $signalSlotDispatcherMock;
-        $result = ProductUtility::getAdditionalButtons($product, $buttons);
-
-        $this->assertEquals($expected, $result);
-    }
 }
